Addressing CR comments
